Add more narration to exitFn
While tracking down the non-usefulness of setting Transport in
logplexc, it was hard to tell what the consequence was of an unadorned
error message.  It was unclear if it was merely advisory or a shutdown
of the client.

This change made it much more clear what the consequences of the log
message was.

Signed-off-by: Daniel Farina <daniel@heroku.com>