Ignore tmp via top level .gitignore
This is in recognition that it's handy to remove the whole tmp
directory in one shot, and it's annoying that its internal .gitignore
sentinel (so that git knows to make the directory) constantly shows up
as a change to be considered for commit.

Tools that use tmp are responsible for making tmp if it doesn't
already exist now, since git will not create it.

Signed-off-by: Daniel Farina <daniel@fdr.io>