Break apart main log processing loop
The main purpose of this is easing testing.  To this end, the
arguments to these procedures are procedure values that are pushed
down into the inner loops, with the intent that test versions of these
procedures can be injected.

exitFn is called when the function needs to exit with an error and
*must* panic.  The intention is it can jump directly to what should-be
the accompanying recover() in the provider of the exitFn.

Signed-off-by: Daniel Farina <daniel@fdr.io>