make testdb: switch to 'git archive' from 'cp -a '
Target REL9_2_2, because 9.3 devel doesn't work (changes to unix
socket directory configuration) and neither does 9.1 (the log hook was
not yet committed).  This was a pain when developing patches against
Postgres' HEAD while also working on pg_logplexcollector.

Change the 'PGSRCDIR' to setting to 'PGSRC', as now all git remotes
are supported as the source of the 'git archive' and 'PGSRCREMOTE'
seems too wordy, and 'PGSRCDIR' is a misnomer.

Also, this removes the utility in issuing 'git clean -fdx', and that
command is dangerous because if the user were to use a source tarball
(instead of git) the 'git clean' command would evaluate in the
contexet of the checked out pg_logplexcollector, cleaning any
untracked files.

Signed-off-by: Daniel Farina <daniel@fdr.io>