Identity to Token database support
Previously, the pg_logfebe identity field was used directly as the
logplex token.  Because the logplex token is deemed sensitive
information, it is desirable to not expose this to the postgresql
configuration, so introduce a layer of indirection in the form of the
token data base, where identities are mapped to real logplex tokens.

Change the 'make testdb' target to make 'tmp' a valid tokendb for
convenience.

Signed-off-by: Daniel Farina <daniel@fdr.io>