TokenDB gives way and becomes ServeDB
The main idea behind this is to support multiple unix sockets to
listen on, as there is a marginal security advantage in that a
compromised LXC container and a poorly configured shared bind mount
(e.g. writable to programs within the container, but shared across
containers) could be used to intercept log traffic otherwise.

As an auxiliary feature, support invalidating old connections and
removal of unix-socket paths -- blindly -- specified in the database,
to prevent unix-socket rebinding errors when re-listening after a
database change.

A better way to handle that might be to accept a directory rather than
a socket path and provide a magically named sentinel file that can
help with both error detection and interlocking.  This is how
PostgreSQL's unix socket handling works.

Signed-off-by: Daniel Farina <daniel@fdr.io>