Un-bitrot serve_db test
Bitrotted from the name field added to the serve struct.  These
records should probably support a dynamic data structure for all the
not-strictly-required stuff.  As-is this was an expeditious hack to
support something similar to Heroku's rendering of colors (which are
used to identify databases on one application) in the log output.

Signed-off-by: Daniel Farina <daniel@fdr.io>